+++ src/UuidFactoryInterface.php
@@ -61,10 +61,13 @@ interface UuidFactoryInterface
      * @param int|null $clockSeq A 14-bit number used to help avoid duplicates
      *     that could arise when the clock is set backwards in time or if the
      *     node ID changes
+     *
+     * @return UuidInterface A UuidInterface instance that represents a
+     *     version 2 UUID
      */
     public function uuid2(
         int $localDomain,
-        ?IntegerValue $localIdentifier,
+        ?IntegerValue $localIdentifier = null,
         ?Hexadecimal $node = null,
         ?int $clockSeq = null
     ): UuidInterface;